+++ examples/state-machine/packages/functions/src/task-two.ts
@@ -1,7 +1,6 @@
-import { withOpenTelemetry } from "./tracing";
+import { ApiHandler } from "sst/node/api";
 
-export const handler = withOpenTelemetry(async (_evt) => {
-  console.log(_evt)
+export const handler = ApiHandler(async (_evt) => {
   return {
     statusCode: 200,
     body: `Hello world. The time is ${new Date().toISOString()}`,